Move test helpers from OmniSharp.Tests into TestUntility
