+++ tests/OmniSharp.Roslyn.CSharp.Tests/NavigateUpDownFacts.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using OmniSharp.Models;
 using OmniSharp.Roslyn.CSharp.Services.Navigation;
-using OmniSharp.Tests;
+using TestUtility;
 using Xunit;
 
 namespace OmniSharp.Roslyn.CSharp.Tests
