+++ tests/OmniSharp.Roslyn.CSharp.Tests/TypeLookupFacts.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 using OmniSharp.Models;
 using OmniSharp.Options;
 using OmniSharp.Roslyn.CSharp.Services.Types;
-using OmniSharp.Tests;
+using TestUtility;
 using Xunit;
 
 namespace OmniSharp.Roslyn.CSharp.Tests
