+++ tests/OmniSharp.Tests/UpdateBufferFilterFacts.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
+using TestUtility;
 using Xunit;
 
 namespace OmniSharp.Tests
